requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Pydantic & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Caching
redis>=5.0.0

# Development
python-multipart>=0.0.6

# Email validation
email-validator>=2.1.0
fastapi-mail>=1.4.1